+++ src/adapter/odata/src/context.rs
@@ -43,18 +43,9 @@ pub(crate) struct ODataServiceContext {
 }
 
 impl ODataServiceContext {
-    pub(crate) fn new(
-        host: &axum::headers::Host,
-        uri: &http::Uri,
-        catalog: Catalog,
-        account_name: Option<AccountName>,
-    ) -> Self {
-        // TODO: Find out scheme the API is served through (e.g. if there is an LB)
-        let scheme = "http";
-        let mut service_base_url = format!("{scheme}://{host}{uri}");
-        if service_base_url.ends_with('/') {
-            service_base_url.pop();
-        }
+    pub(crate) fn new(catalog: Catalog, account_name: Option<AccountName>) -> Self {
+        let config = catalog.get_one::<ServerUrlConfig>().unwrap();
+        let service_base_url = config.protocols.odata_base_url();
 
         Self {
             catalog,
@@ -118,23 +109,12 @@ pub(crate) struct ODataCollectionContext {
 
 impl ODataCollectionContext {
     pub(crate) fn new(
-        host: &axum::headers::Host,
-        uri: &http::Uri,
         catalog: Catalog,
         dataset_handle: DatasetHandle,
         dataset: Arc<dyn Dataset>,
     ) -> Self {
-        let scheme = std::env::var("KAMU_PROTOCOL_SCHEME").unwrap_or_else(|_| String::from("http"));
-        let (base_path, _) = uri
-            .path_and_query()
-            .unwrap()
-            .path()
-            .rsplit_once('/')
-            .unwrap();
-        let mut service_base_url = format!("{scheme}://{host}{base_path}");
-        if service_base_url.ends_with('/') {
-            service_base_url.pop();
-        }
+        let config = catalog.get_one::<ServerUrlConfig>().unwrap();
+        let service_base_url = config.protocols.odata_base_url();
 
         Self {
             catalog,