+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added GrapqQL Dataset Endpoints object
 ### Changed
 - REST API: `/ingest` endpoint will return HTTP 400 error when data cannot be read correctly
 - Improved API token generation command